APIs/security/views.py
from rest_framework.decorators import api_view, permission_classes
from rest_framework.permissions import IsAuthenticated
from rest_framework.response import Response
from rest_framework import status
from django.core.files.storage import default_storage
import hashlib
import magic
from datetime import datetime

@api_view(['POST'])
@permission_classes([IsAuthenticated])
def scan_file(request):
    """Scan uploaded file for malware"""
    
    if 'file' not in request.FILES:
        return Response({'error': 'No file provided'}, status=status.HTTP_400_BAD_REQUEST)
    
    uploaded_file = request.FILES['file']
    
    # Basic file analysis
    file_hash = hashlib.sha256(uploaded_file.read()).hexdigest()
    uploaded_file.seek(0)  # Reset file pointer
    
    try:
        file_type = magic.from_buffer(uploaded_file.read(1024), mime=True)
        uploaded_file.seek(0)
    except:
        file_type = 'unknown'
    
    # Mock scan results - Replace with real scanning logic
    scan_result = {
        'file_name': uploaded_file.name,
        'file_size': uploaded_file.size,
        'file_hash': file_hash,
        'file_type': file_type,
        'scan_time': datetime.now().isoformat(),
        'threat_detected': False,
        'threat_type': None,
        'risk_level': 'Low',
        'scan_engine': 'Jorise Scanner v2.0',
        'details': {
            'signatures_matched': 0,
            'suspicious_behaviors': [],
            'reputation_score': 85
        }
    }
    
    return Response(scan_result)

@api_view(['POST'])
@permission_classes([IsAuthenticated])
def scan_url(request):
    """Scan URL for malicious content"""
    
    url = request.data.get('url')
    if not url:
        return Response({'error': 'No URL provided'}, status=status.HTTP_400_BAD_REQUEST)
    
    # Mock URL scan - Replace with real implementation
    scan_result = {
        'url': url,
        'scan_time': datetime.now().isoformat(),
        'threat_detected': False,
        'categories': ['Technology', 'Software'],
        'reputation_score': 92,
        'risk_level': 'Low',
        'details': {
            'phishing_detected': False,
            'malware_hosting': False,
            'suspicious_redirects': False,
            'ssl_certificate': 'Valid'
        }
    }
    
    return Response(scan_result)

@api_view(['GET'])
@permission_classes([IsAuthenticated])
def scan_history(request):
    """Get user's scan history"""
    
    # Mock data - Replace with database queries
    history = [
        {
            'id': 1,
            'file_name': 'document.pdf',
            'scan_time': '2025-10-09T10:30:00Z',
            'result': 'Clean',
            'threat_detected': False
        },
        {
            'id': 2,
            'file_name': 'suspicious.exe',
            'scan_time': '2025-10-09T09:15:00Z',
            'result': 'Threat Detected',
            'threat_detected': True
        }
    ]
    
    return Response(history)

@api_view(['GET'])
@permission_classes([IsAuthenticated])
def threat_list(request):
    """Get list of detected threats"""
    
    threats = [
        {
            'id': 1,
            'threat_name': 'Trojan.Generic.KD.12345',
            'severity': 'High',
            'detected_time': '2025-10-09T08:45:00Z',
            'source': '192.168.1.105',
            'status': 'Quarantined'
        }
    ]
    
    return Response(threats)

@api_view(['GET'])
@permission_classes([IsAuthenticated])
def threat_detail(request, pk):
    """Get detailed threat information"""
    
    threat_detail = {
        'id': pk,
        'threat_name': 'Trojan.Generic.KD.12345',
        'severity': 'High',
        'detected_time': '2025-10-09T08:45:00Z',
        'source': '192.168.1.105',
        'status': 'Quarantined',
        'description': 'Generic trojan detection',
        'mitigation_steps': [
            'File has been quarantined',
            'System scan recommended',
            'Update antivirus definitions'
        ]
    }
    
    return Response(threat_detail)

@api_view(['GET'])
@permission_classes([IsAuthenticated])
def protection_status(request):
    """Get current protection status"""
    
    status_data = {
        'firewall': {'enabled': True, 'status': 'Active'},
        'antivirus': {'enabled': True, 'status': 'Up to date'},
        'web_protection': {'enabled': True, 'status': 'Active'},
        'email_protection': {'enabled': True, 'status': 'Active'},
        'last_update': '2025-10-09T06:00:00Z'
    }
    
    return Response(status_data)

@api_view(['GET'])
@permission_classes([IsAuthenticated])
def firewall_rules(request):
    """Get firewall rules"""
    
    rules = [
        {
            'id': 1,
            'name': 'Block suspicious IPs',
            'action': 'Block',
            'source': 'Threat Intelligence',
            'enabled': True
        },
        {
            'id': 2,
            'name': 'Allow internal network',
            'action': 'Allow',
            'source': '192.168.0.0/16',
            'enabled': True
        }
    ]
    
    return Response(rules)