requirements.txt
# Jorise v2 - Enterprise Security Operations Center
# Python Dependencies

# Django & Web Framework
Django>=4.2,<5.1
djangorestframework>=3.14,<4.0
django-cors-headers>=4.0,<5.0
gunicorn>=21.0,<22.0
whitenoise>=6.6,<7.0

# Database
psycopg2-binary>=2.9,<3.0
dj-database-url>=2.0,<3.0

# Authentication & Security
djangorestframework-simplejwt>=5.3,<6.0
django-allauth>=0.57,<1.0
cryptography>=41.0,<43.0

# AI & Machine Learning
google-generativeai>=0.3,<1.0

# Task Queue
celery>=5.3,<6.0
redis>=5.0,<6.0
django-celery-beat>=2.5,<3.0

# Monitoring & Logging
python-json-logger>=2.0,<3.0

# PDF Generation
reportlab>=4.0,<5.0
Pillow>=10.0,<11.0

# Payment Processing
stripe>=7.0,<8.0

# Utils
python-decouple>=3.8,<4.0
python-dotenv>=1.0,<2.0
pytz>=2023.3
python-dateutil>=2.8,<3.0
requests>=2.31,<3.0

# PCAP & Network Analysis
scapy>=2.5,<3.0

# Machine Learning & Data Science
scikit-learn>=1.3,<2.0
pandas>=2.0,<3.0
numpy>=1.24,<2.0
joblib>=1.3,<2.0
matplotlib>=3.7,<4.0

